Flyttet ut ytelsePerson til egen util og util-test filer. Forbedret noen av testene etter samarbeid med Eivind.
