+++ src/test/kotlin/no/nav/familie/ba/sak/behandling/resultat/YtelsePersonUtilsTest.kt
@@ -1,0 +1,588 @@
+package no.nav.familie.ba.sak.behandling.resultat
+
+import no.nav.familie.ba.sak.beregning.domene.YtelseType
+import no.nav.familie.ba.sak.common.forrigeMåned
+import no.nav.familie.ba.sak.common.inneværendeMåned
+import no.nav.familie.ba.sak.common.lagAndelTilkjentYtelse
+import no.nav.familie.ba.sak.common.lagSøknadDTO
+import no.nav.familie.ba.sak.common.tilfeldigPerson
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertFalse
+import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.Test
+
+class YtelsePersonUtilsTest {
+
+    val søker = tilfeldigPerson()
+    val barn1 = tilfeldigPerson()
+    val barn2 = tilfeldigPerson()
+
+    @Test
+    fun `Skal kun finne søknadsytelsePersoner`() {
+        val søknadDTO = lagSøknadDTO(
+                søkerIdent = søker.personIdent.ident,
+                barnasIdenter = listOf(barn1.personIdent.ident)
+        )
+
+        val ytelsePersoner = YtelsePersonUtils.utledKrav(
+                søknadDTO = søknadDTO,
+                forrigeAndelerTilkjentYtelse = emptyList()
+        )
+
+        assertEquals(1, ytelsePersoner.size)
+        assertEquals(barn1.personIdent.ident, ytelsePersoner.first().personIdent)
+        assertEquals(YtelseType.ORDINÆR_BARNETRYGD, ytelsePersoner.first().ytelseType)
+        assertTrue(ytelsePersoner.first().erFramstiltKravForINåværendeBehandling)
+    }
+
+    @Test
+    fun `Skal kun finne endringsytelsePersoner`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                       "2020-01",
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val ytelsePersoner = YtelsePersonUtils.utledKrav(
+                søknadDTO = null,
+                forrigeAndelerTilkjentYtelse = listOf(forrigeAndelBarn1)
+        )
+
+        assertEquals(1, ytelsePersoner.size)
+        assertEquals(barn1.personIdent.ident, ytelsePersoner.first().personIdent)
+        assertEquals(YtelseType.ORDINÆR_BARNETRYGD, ytelsePersoner.first().ytelseType)
+        assertFalse(ytelsePersoner.first().erFramstiltKravForINåværendeBehandling)
+    }
+
+    @Test
+    fun `Skal finne 2 endringsytelsePersoner på samme barn`() {
+        val forrigeAndelBarn1Ordinær = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                              "2020-01",
+                                                              YtelseType.ORDINÆR_BARNETRYGD,
+                                                              1054,
+                                                              person = barn1)
+
+        val forrigeAndelBarn1Utvidet = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                              "2020-01",
+                                                              YtelseType.UTVIDET_BARNETRYGD,
+                                                              1054,
+                                                              person = barn1)
+
+        val ytelsePersoner = YtelsePersonUtils.utledKrav(
+                søknadDTO = null,
+                forrigeAndelerTilkjentYtelse = listOf(forrigeAndelBarn1Ordinær, forrigeAndelBarn1Utvidet)
+        )
+
+        assertEquals(2, ytelsePersoner.size)
+        assertTrue(ytelsePersoner.any { it.personIdent == barn1.personIdent.ident && it.ytelseType == YtelseType.ORDINÆR_BARNETRYGD && !it.erFramstiltKravForINåværendeBehandling })
+        assertTrue(ytelsePersoner.any { it.personIdent == barn1.personIdent.ident && it.ytelseType == YtelseType.UTVIDET_BARNETRYGD && !it.erFramstiltKravForINåværendeBehandling })
+    }
+
+    @Test
+    fun `Skal finne 1 av 2 endringsytelsePersoner og 1 søknadsytelsePersoner`() {
+        val søknadDTO = lagSøknadDTO(
+                søkerIdent = søker.personIdent.ident,
+                barnasIdenter = listOf(barn1.personIdent.ident)
+        )
+
+        val forrigeAndelBarn1Ordinær = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                              "2020-01",
+                                                              YtelseType.ORDINÆR_BARNETRYGD,
+                                                              1054,
+                                                              person = barn1)
+
+        val forrigeAndelBarn1Utvidet = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                              "2020-01",
+                                                              YtelseType.UTVIDET_BARNETRYGD,
+                                                              1054,
+                                                              person = barn1)
+
+        val ytelsePersoner = YtelsePersonUtils.utledKrav(
+                søknadDTO = søknadDTO,
+                forrigeAndelerTilkjentYtelse = listOf(forrigeAndelBarn1Ordinær, forrigeAndelBarn1Utvidet)
+        )
+
+        assertEquals(2, ytelsePersoner.size)
+        assertTrue(ytelsePersoner.any { it.personIdent == barn1.personIdent.ident && it.ytelseType == YtelseType.ORDINÆR_BARNETRYGD && it.erFramstiltKravForINåværendeBehandling })
+        assertTrue(ytelsePersoner.any { it.personIdent == barn1.personIdent.ident && it.ytelseType == YtelseType.UTVIDET_BARNETRYGD && !it.erFramstiltKravForINåværendeBehandling })
+    }
+
+
+    // Tester for ytelse person resultater
+    @Test
+    fun `Skal utelede innvilget første gang barnet blir vurdert`() {
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                inneværendeMåned().plusYears(2).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = emptyList(),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1)
+        )
+
+        assertEquals(1, ytelsePersonerMedResultat.size)
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utelede innvilget og redusert første gang barnet blir vurdert med kun etterbetaling`() {
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                inneværendeMåned().plusYears(2).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = emptyList(),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1)
+        )
+
+        assertEquals(1, ytelsePersonerMedResultat.size)
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede innvilget første gang barn nr2 blir vurdert i en revurdering`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(5).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn2 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                inneværendeMåned().plusYears(1).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn2)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+                YtelsePerson(
+                        personIdent = barn2.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1,
+                                                                                                        andelBarn2)
+        )
+
+        assertEquals(2, ytelsePersonerMedResultat.size)
+        assertEquals(emptySet<YtelsePersonResultat>(),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn2.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede innvilget og redusert første gang barn nr2 blir vurdert i en revurdering med kun etterbetaling`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(5).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn2 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(3).toString(),
+                                                inneværendeMåned().minusYears(1).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn2)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+                YtelsePerson(
+                        personIdent = barn2.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1,
+                                                                                                        andelBarn2)
+        )
+
+        assertEquals(2, ytelsePersonerMedResultat.size)
+        assertEquals(emptySet<YtelsePersonResultat>(),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET, YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn2.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede redusert når barn nr1 blir redusert i en revurdering`() {
+        val reduksjonsmåned = inneværendeMåned()
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(5).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(5).toString(),
+                                                reduksjonsmåned.toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1)
+        )
+
+        assertEquals(1, ytelsePersonerMedResultat.size)
+        assertEquals(setOf(YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+        assertEquals(reduksjonsmåned.plusMonths(1),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.periodeStartForRentOpphør)
+    }
+
+    @Test
+    fun `Skal utelede avslag første gang ytelsePersoneret for barn fremstilles`() {
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = emptyList(),
+                                                                                                andelerTilkjentYtelse = emptyList()
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.AVSLÅTT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede avslag på søknad for nytt barn i revurdering`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+                YtelsePerson(
+                        personIdent = barn2.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1)
+        )
+
+        assertEquals(emptySet<YtelsePersonResultat>(),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+        assertEquals(setOf(YtelsePersonResultat.AVSLÅTT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn2.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede redusert på revurdering hvor alle andeler er annulert`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf()
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede fortsatt innvilget på årlig kontroll uten endringer`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1)
+        )
+
+        assertEquals(emptySet<YtelsePersonResultat>(),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede endring på årlig kontroll med liten endring tilbake i tid`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                inneværendeMåned().minusMonths(12).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val andel2Barn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusMonths(10).toString(),
+                                                 inneværendeMåned().plusMonths(12).toString(),
+                                                 YtelseType.ORDINÆR_BARNETRYGD,
+                                                 1054,
+                                                 person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1,
+                                                                                                        andel2Barn1)
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.ENDRET),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+    @Test
+    fun `Skal utlede endring og opphør på årlig kontroll med liten endring tilbake i tid og opphør`() {
+        val reduksjonsmåned = inneværendeMåned().forrigeMåned()
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                inneværendeMåned().minusMonths(12).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val andel2Barn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusMonths(10).toString(),
+                                                 reduksjonsmåned.toString(),
+                                                 YtelseType.ORDINÆR_BARNETRYGD,
+                                                 1054,
+                                                 person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1,
+                                                                                                        andel2Barn1)
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+        assertEquals(null,
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.periodeStartForRentOpphør)
+    }
+
+    @Test
+    fun `Skal utelede innvilget andre gang ytelsePersoneret for barn fremstilles`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       "2019-01",
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                "2020-01",
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1)
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET, YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+
+
+    @Test
+    fun `Skal utlede innvilget på søknad for nytt barn i revurdering`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        val andelBarn2 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(2).toString(),
+                                                inneværendeMåned().plusMonths(12).toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn2)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+                YtelsePerson(
+                        personIdent = barn2.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = true
+                )
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1,
+                                                                                                        andelBarn2)
+        )
+
+        assertEquals(emptySet<YtelsePersonResultat>(),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+
+        assertEquals(setOf(YtelsePersonResultat.INNVILGET),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn2.personIdent.ident }?.resultater)
+    }
+
+
+    @Test
+    fun `Skal utlede opphør på endring for barn i revurdering`() {
+        val forrigeAndelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                       inneværendeMåned().plusMonths(12).toString(),
+                                                       YtelseType.ORDINÆR_BARNETRYGD,
+                                                       1054,
+                                                       person = barn1)
+
+        // Opphør ytelsen for barnet 1 måned tilbake i tid
+        val andelBarn1 = lagAndelTilkjentYtelse(inneværendeMåned().minusYears(4).toString(),
+                                                inneværendeMåned().toString(),
+                                                YtelseType.ORDINÆR_BARNETRYGD,
+                                                1054,
+                                                person = barn1)
+
+        val ytelsePersoner = listOf(
+                YtelsePerson(
+                        personIdent = barn1.personIdent.ident,
+                        ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
+                        erFramstiltKravForINåværendeBehandling = false
+                ),
+        )
+
+        val ytelsePersonerMedResultat = YtelsePersonUtils.utledYtelsePersonerMedResultat(ytelsePersoner = ytelsePersoner,
+                                                                                                forrigeAndelerTilkjentYtelse = listOf(
+                                                                                                        forrigeAndelBarn1),
+                                                                                                andelerTilkjentYtelse = listOf(
+                                                                                                        andelBarn1)
+        )
+
+        assertEquals(setOf(YtelsePersonResultat.OPPHØRT),
+                                ytelsePersonerMedResultat.find { it.personIdent == barn1.personIdent.ident }?.resultater)
+    }
+}
